+++ src/extension.ts
@@ -7,49 +7,63 @@ let timerStatus: vscode.StatusBarItem;
 export function activate(context: vscode.ExtensionContext) {
 	console.log("Your \"body.move()\" extension is now active.");
 	let waitTime = 0; // to be set by user via SetReminderTime() function
+	let timer: any; // a timeout to be set by InitializeReminder()
+	let messages = ["Time to get up and move!", "Go take a walk!", "You should stand up and strech!", "Go get some fresh air!",
+									"Go do some exercise!"];
+
+	// status bar
 	timerStatus = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right, 100);
 	timerStatus.color = '#fff';
 	context.subscriptions.push(timerStatus);
 
+	//user commands to be run from Command Pallette
+	const setMoveTimer = 'body-move.setMoveTimer';
+	context.subscriptions.push(vscode.commands.registerCommand(setMoveTimer, SetReminderTime));
+	const clearMoveTimer = 'body-move.clearMoveTimer';
+	context.subscriptions.push(vscode.commands.registerCommand(clearMoveTimer, ClearReminder));
+
 	// ask the user to enter how many minutes to delay the reminder
 	function SetReminderTime() {
 		vscode.window.showInputBox({placeHolder:
-			"How many minutes should I wait to remind you to get up and move?"}).then(function (answer) { 
-				if(answer){ 
-				let timeSelection = parseInt(answer);
+			"How many minutes from now would you like to be reminded to get up and move?"}).then(function (answer) { 
+			if(answer){ 
+				let timeEntry = parseInt(answer);
 				// validate the user's input
-				if(timeSelection > 0) {
-					vscode.window.showInformationMessage("I\'ll remind you to get up in " + timeSelection + " minute(s).");
-					waitTime = timeSelection * 60000;
+				if(timeEntry > 0) {
+					vscode.window.showInformationMessage("body.move() reminder set for " + timeEntry + " minute(s).");
+					waitTime = timeEntry * 60000;
 					InitializeReminder();
-				}
+				} // if
 				else {
 					vscode.window.showErrorMessage("Please enter a valid integer, greater than zero.");
 					SetReminderTime();
-				}
+				} // else
 			} // outer if
-		}); // showInputBox()
+		}); // .then(function())
 	} //SetReminderTime()
 
 	SetReminderTime();
 
-	// trim extra seconds and set reminder to fire at the top of the the minute (currently only an issue on first run)
-	// intending to add a way that user can adjust the interval via command
+	function ClearReminder() {
+		clearTimeout(timer);
+		vscode.window.showInformationMessage("body.move() reminder cleared.");
+		timerStatus.text = "body.move() not set.";
+	} // ClearReminder()
+
+	// trim extra seconds and set reminder to fire at the top of the the minute
 	function InitializeReminder() {
 		let date = new Date();
 		let curSeconds = date.getSeconds() * 1000;
 		let intervalTime = waitTime - curSeconds;
-		let timer = setTimeout(ShowReminder, intervalTime);
+		timer = setTimeout(ShowReminder, intervalTime);		
 		UpdateStatusBar();
 	} // InitializeReminder()
 
 	// fire off the reminder via Information Message
 	function ShowReminder() {
-		let messages = ["Time to get up and move!", "Go take a walk!", "You should stand up and strech!", "Go get some fresh air!",
-									"Go do some exercise!"];
-		// randomly select a message to display
-		let reminderMessage = messages[Math.floor(Math.random() * messages.length)];
-		vscode.window.showInformationMessage(reminderMessage);
+		// randomly select a message to display from messages[]
+		let message = messages[Math.floor(Math.random() * messages.length)];
+		vscode.window.showInformationMessage(message);
 		// queue the next reminder
 		InitializeReminder();
 	} // Move()
@@ -58,6 +72,7 @@ export function activate(context: vscode.ExtensionContext) {
 	function UpdateStatusBar() {
 		let hours = "";
 		let minutes = "";
+		let meridiem = "";
 		let date = new Date();
 		date.setTime(date.getTime() + waitTime); // set the reminder time to be displayed in status bar
 		hours = date.getHours().toString();
@@ -74,21 +89,23 @@ export function activate(context: vscode.ExtensionContext) {
 		if(date.getHours() % 24 > 12) {
 			date.setHours(date.getHours() - 12);
 			hours = date.getHours().toString();
-			timerStatus.text = "body.move() at " + hours + ":" + minutes + " PM.";
-		}
-		// if midnight, change 00 to 12
-		else if(date.getHours() === 0) {
-			date.setHours(12);
-			hours = date.getHours().toString();
-			timerStatus.text = "body.move() at " + hours + ":" + minutes + " AM.";
+			meridiem = " PM.";
 		}
 		else {
-			timerStatus.text = "body.move() at " + hours + ":" + minutes + " AM.";
+			meridiem = " AM.";
+		}
+
+		// if midnight, change 00 to 12
+		if(date.getHours() === 0) {
+			date.setHours(12);
+			hours = date.getHours().toString();
 		}
 		
+		// set the text that will show in the status bar
+		timerStatus.text = "body.move() at " + hours + ":" + minutes + meridiem;
+				
 		// print the message to status bar
 		timerStatus.show();
-
 	} // UpdateStatusBar()
 
 } // activate()