+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 All notable changes to the "body-move" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+### 0.1.2
+04-25-2019
+- Added setMoveTimer command to adjust the reminder interval
+- Added clearMoveTimer command to clear the reminder interval, thus disabling reminders
 
-## [Unreleased]
+### 0.0.2
+04-23-2019
+- Shortened the message in the status bar.
 
-- Initial release
+### 0.0.1
+04-23-2019
+- Initial release of body-move VS Code extension.